// lib/schemas/auth.ts
import { z } from 'zod';

export const loginSchema = z.object({
  email: z.string().email('Invalid email address'),
  password: z.string().min(1, 'Password is required'),
});
const registerSchema = z.object({
    name: z.string().min(2 , 'Name must be at least 2 characters'),
    email: z.string().email('Invalid email address'),
    password: z.string().min(6 , 'Password must be at least 6 characters'),
    confirmPassword: z.string().min(6)
}).refine(data => data.password === data.confirmPassword, {
    message: "Passwords do not match",
    path: ["confirmPassword"]
});

export type LoginInput = z.infer<typeof loginSchema>;
export type RegisterInput = z.infer<typeof registerSchema>;